+++ src/components/event-edit.js
@@ -1,4 +1,4 @@
-import {getRandomBool} from '../utils.js';
+import {getRandomBool} from '../utils/common.js';
 import {formatDate, formatTime} from '../utils/format.js';
 import {EVENT_TYPES} from '../const.js';
 import {Destinations} from '../mock/mock.js';
